+++ R/select.R
@@ -70,8 +70,8 @@
 #' @examples
 #' \donttest{
 #' data("sim_binary", package = "gbtmkit")
-#' spec <- gbtm_spec(sim_binary, c("y1","y2","y3","y4"),
-#'                   c("t1","t2","t3","t4"), id = "id", family = "binomial")
+#' spec <- gbtm_spec(sim_binary, paste0("y", 1:10),
+#'                   paste0("t", 1:10), id = "id", family = "binomial")
 #' if (requireNamespace("trajeR", quietly = TRUE))
 #'   select_algorithm(spec, n_groups = 4, degrees = rep(1, 4))
 #' }
@@ -138,10 +138,10 @@ select_algorithm <- function(spec,
 #' @examples
 #' \donttest{
 #' data("sim_binary", package = "gbtmkit")
-#' spec <- gbtm_spec(sim_binary, c("y1","y2","y3","y4"),
-#'                   c("t1","t2","t3","t4"), id = "id", family = "binomial")
+#' spec <- gbtm_spec(sim_binary, paste0("y", 1:10),
+#'                   paste0("t", 1:10), id = "id", family = "binomial")
 #' if (requireNamespace("trajeR", quietly = TRUE))
-#'   select_n_groups(spec, candidates = 2:5, degree = 1)
+#'   select_n_groups(spec, candidates = 2:5, degree = 2)
 #' }
 #' @export
 select_n_groups <- function(spec,